Move StackingContext to a separate file
